Move `File` and `Fileset` types to root `file` module
