+++ packages/ploys/src/file/fileset.rs
@@ -5,6 +5,8 @@ use std::path::{Path, PathBuf};
 
 use crate::package::{Lockfile, Package, PackageKind};
 
+use super::File;
+
 /// A collection of files.
 #[derive(Clone, Debug, Default)]
 pub struct Fileset {
@@ -155,66 +157,5 @@ where
                 .map(|file| (file.path().to_owned(), Some(file)))
                 .collect(),
         }
-    }
-}
-
-/// A file in one of a number of formats.
-#[derive(Clone, Debug)]
-pub enum File {
-    Package(Package),
-    Lockfile(Lockfile),
-}
-
-impl File {
-    /// Gets the file path.
-    pub fn path(&self) -> &Path {
-        match self {
-            Self::Package(package) => package.path(),
-            Self::Lockfile(lockfile) => lockfile.kind().lockfile_name().expect("path"),
-        }
-    }
-
-    /// Gets the file as a package.
-    pub fn as_package(&self) -> Option<&Package> {
-        match self {
-            Self::Package(package) => Some(package),
-            _ => None,
-        }
-    }
-
-    /// Gets the file as a mutable package.
-    pub fn as_package_mut(&mut self) -> Option<&mut Package> {
-        match self {
-            Self::Package(package) => Some(package),
-            _ => None,
-        }
-    }
-
-    /// Gets the file as a lockfile.
-    pub fn as_lockfile(&self) -> Option<&Lockfile> {
-        match self {
-            Self::Lockfile(lockfile) => Some(lockfile),
-            _ => None,
-        }
-    }
-
-    /// Gets the file as a mutable lockfile.
-    pub fn as_lockfile_mut(&mut self) -> Option<&mut Lockfile> {
-        match self {
-            Self::Lockfile(lockfile) => Some(lockfile),
-            _ => None,
-        }
-    }
-}
-
-impl From<Package> for File {
-    fn from(value: Package) -> Self {
-        Self::Package(value)
-    }
-}
-
-impl From<Lockfile> for File {
-    fn from(value: Lockfile) -> Self {
-        Self::Lockfile(value)
     }
 }
