+++ packages/ploys/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod changelog;
+pub mod file;
 pub mod package;
 pub mod project;
 pub mod repository;
